feat: add videos to mirage server
