+++ src/mirage/models.ts
@@ -1,0 +1,9 @@
+import { Model } from "miragejs";
+
+const models = {
+    video: Model.extend({}),
+    user: Model.extend({}),
+    plugin: Model.extend({}),
+};
+
+export default models;
